+++ spec/classes/init_spec.rb
@@ -1,14 +1,12 @@
 require 'spec_helper'
 describe 'gogs' do
-  context 'with default values for all parameters' do
-    let(:facts) do
-      {
-          :osfamily => 'Debian',
-          :operatingsystem => 'Ubuntu',
-          :architecture => 'amd46'
-      }
+  on_supported_os.each do |os, facts|
+    context "on #{os} with defaults" do
+      let(:facts) do
+        facts
+      end
+      let(:pre_condition) { "define puppetstats($enabled) {}" }
+      it {should contain_class('gogs')}
     end
-    let(:pre_condition) { "define puppetstats($enabled) {}" }
-    it {should contain_class('gogs')}
   end
 end